requirements.txt
AnilistPython==0.1.3
anyio==3.6.2
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
CacheControl==0.12.11
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cinemagoer==2022.12.27
click==8.1.3
cryptography==40.0.2
cssselect==1.2.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
firebase-admin==6.1.0
Flask==2.2.3
Flask-Session==0.4.0
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
hyperframe==6.0.1
idna==3.4
IMDbPY==2022.7.9
importlib-metadata==6.5.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jws==0.1.3
lxml==4.9.2
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.2
oauth2client==3.0.0
packaging==23.1
parse==1.19.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.4.3
pyee==8.2.2
PyJWT==2.6.0
PyMovieDb==0.0.7
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
Pyrebase==3.0.27
pytest==7.3.1
python-anilist==1.0.9
python-dotenv==1.0.0
python-jwt==2.0.1
requests==2.28.2
requests-html==0.10.0
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.9
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
w3lib==2.1.1
websockets==10.4
Werkzeug==2.2.3
zipp==3.15.0
